requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
shortuuid==1.0.11
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
